refactor: better error handling
the debug decorator has been removed and replaced by @tc.typing to
verify the correct type on runtime. It may a performance issue but
for this scenario it is better to have correct types.

if there is a validation error, there is now created a file with with
the error.

removed the parent parameter it was nowhere used.